+++ src/app/features/admin/admin.routing.module.ts
@@ -3,8 +3,8 @@ import { RouterModule, Routes } from '@angular/router';
 import { GetDatesResolver } from '@core/resolvers/admin/local-authorities-return/get-dates.resolver';
 
 import { LocalAuthoritiesReturnComponent } from './local-authorities-return/local-authorities-return.component';
+import { SetDatesComponent } from './local-authorities-return/set-dates/set-dates.component';
 import { SearchComponent } from './search/search.component';
-import { SetDatesComponent } from './set-dates/set-dates.component';
 
 const routes: Routes = [
   // {
